se movió al main activity el textField nombre
