+++ app/src/main/java/ar/edu/unlam/mobile2/ui/MainActivity.kt
@@ -1,6 +1,7 @@
 package ar.edu.unlam.mobile2.ui
 
 import android.content.Intent
+import android.media.MediaPlayer
 import android.os.Bundle
 import android.util.Log
 import android.view.Menu
@@ -8,17 +9,31 @@ import android.view.MenuItem
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Button
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.input.TextFieldValue
+import androidx.compose.ui.unit.dp
 import androidx.navigation.ui.AppBarConfiguration
 import ar.edu.unlam.mobile2.R
 import ar.edu.unlam.mobile2.databinding.ActivityMainBinding
@@ -47,6 +62,9 @@ class MainActivity : /*AppCompatActivity()*/ ComponentActivity() {
                     verticalArrangement = Arrangement.SpaceAround,
                     horizontalAlignment = Alignment.CenterHorizontally
                 ) {
+                    nameEditText()
+
+
                     Button(onClick = {
                         irAJuego()
                         onStop()
@@ -65,14 +83,19 @@ class MainActivity : /*AppCompatActivity()*/ ComponentActivity() {
                     }) {
                         Texto("Historial")
                     }
+
+
                 }
             }
         }
     }
 
+
+
     override fun onStart() {
         super.onStart()
         Log.i("MainActivity", "onStart")
+
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
@@ -116,4 +139,28 @@ class MainActivity : /*AppCompatActivity()*/ ComponentActivity() {
     fun Texto(texto: String) {
         Text(texto, color = Color.White)
     }
+
+    @OptIn(ExperimentalMaterial3Api::class)
+    @Composable
+    fun nameEditText(){
+
+        var text by remember { mutableStateOf(TextFieldValue("")) }
+
+        OutlinedTextField(
+            value = text,
+            onValueChange = {text = it},
+            label = { Text(text = "Nombre")},
+            placeholder = {Text(text = "Nombre")},
+            colors = TextFieldDefaults.outlinedTextFieldColors(
+                cursorColor = Color.Red
+            ),
+            singleLine = true,
+            modifier = Modifier
+                .padding(top = 20.dp)
+
+        )
+
+    }
+
+
 }
